Updated HSQLDB to 2.6.1
- Updated HSQLDB to v2.6.1
- Updated several tests accordingly
- Moved some tests to more suitable test suites
