+++ server/pom.xml
@@ -877,7 +877,7 @@
       <dependency>
         <groupId>org.hsqldb</groupId>
         <artifactId>hsqldb</artifactId>
-        <version>2.3.3</version>
+        <version>2.6.1</version>
       </dependency>
       <dependency>
         <groupId>com.google.inject.extensions</groupId>
